feat: 비디오 타입 분리
