+++ types/video.ts
@@ -1,0 +1,6 @@
+export type VideoInfo = {
+	userId: string;
+	userName: string;
+	stream: MediaStream;
+	isOwner: boolean;
+};
